requirements.txt
langchain
chromadb
boto3
pandas
awscli
python-dotenv
streamlit
langchain_aws
faiss-cpu